Move permission name translation into a trait
